# grpc-gateway-eg/proto/notes/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Notes",
    "version": "1.0",
    "contact": {
      "name": "bbengfort",
      "url": "https://github.com/bbengfort/notes",
      "email": "info@bengfort.com"
    },
    "license": {
      "name": "BSD 3-Clause License",
      "url": "https://github.com/bbengfort/notes/LICENSE"
    }
  },
  "tags": [
    {
      "name": "NoteService"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/notes": {
      "get": {
        "operationId": "NoteService_Fetch",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Notebook"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uint64"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "private",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "NoteService"
        ]
      },
      "post": {
        "operationId": "NoteService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Notebook"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Note"
            }
          }
        ],
        "tags": [
          "NoteService"
        ]
      }
    },
    "/v1/identity/{identity.type}/register-by-link/cancel": {
      "post": {
        "operationId": "NoteService_CancelRegisterIdentityByInstructionLink",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Notebook"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "identity.type",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NoteServiceCancelRegisterIdentityByInstructionLinkBody"
            }
          }
        ],
        "tags": [
          "NoteService"
        ]
      }
    }
  },
  "definitions": {
    "NoteServiceCancelRegisterIdentityByInstructionLinkBody": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "description": "The provisioned User ID."
        },
        "companyId": {
          "type": "string",
          "description": "The company ID where the corporate phone number belongs."
        },
        "identity": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "objectId": {
              "type": "string",
              "title": "database entity id"
            },
            "companyId": {
              "type": "string"
            },
            "verificationState": {
              "type": "string"
            }
          },
          "description": "Contains 2 properties: `id` (string, required) and `attributes` (array of objects). Property `id` is the corporate phone number. Property `attributes` contains `com.leapxpert.official.account.phone.number` for key `name`, `STRING` for key `type` and corporate phone number for key `values`.",
          "title": "Contains 2 properties: `id` (string, required) and `attributes` (array of objects). Property `id` is the corporate phone number. Property `attributes` contains `com.leapxpert.official.account.phone.number` for key `name`, `STRING` for key `type` and corporate phone number for key `values`."
        },
        "corporateAccountId": {
          "type": "string",
          "description": "The corporate phone number ID."
        }
      },
      "required": [
        "userId",
        "companyId"
      ]
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1FederatedIdentity": {
      "type": "object",
      "properties": {
        "type": {
          "type": "integer",
          "format": "int64"
        },
        "id": {
          "type": "string"
        },
        "objectId": {
          "type": "string",
          "title": "database entity id"
        },
        "companyId": {
          "type": "string"
        },
        "verificationState": {
          "type": "string"
        }
      }
    },
    "v1Note": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "timestamp": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "private": {
          "type": "boolean"
        }
      }
    },
    "v1Notebook": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/v1Error"
        },
        "notes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Note"
          }
        }
      }
    }
  }
}